Split far_transfer.rs into subdirectory modules
Split the 1010-line far_transfer.rs into a far_transfer/ directory with
focused sub-modules to improve maintainability:

- mod.rs (66 lines): re-exports and FarTransferPlugin
- types.rs (259 lines): constants, FarBonusType, FarTransferState, helpers, Saveable
- systems.rs (193 lines): update_far_bonuses, update_far_transfers systems
- tests.rs (368 lines): unit tests for types and helpers
- tests_tdr.rs (134 lines): TDR, stat district, and saveable tests

All public items are re-exported from mod.rs so external callers are
unaffected. No functional changes.

Part of #1389
